% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_errors.R
\name{calculate_errors}
\alias{calculate_errors}
\title{Calculate errors in matrix values based on sample sizes.}
\usage{
calculate_errors(mat_U, mat_F, sample_size, type = "sem")
}
\arguments{
\item{mat_U}{A matrix of proportions.}

\item{mat_F}{A matrix of proportions.}

\item{sample_size}{A positive number indicating the sample size.}

\item{type}{A character string indicating the type of error to calculate.
Must be one of "sem" (standard error), or "CI95" (95\% confidence interval).}
}
\value{
A list containing the error estimates. If type is "\code{CI95}", the list
contains the upper and lower confidence intervals for both matrices
(\code{mat_U_upperCI}, \code{mat_U_lowerCI}, \code{mat_F_upperCI}, \code{mat_F_lowerCI}).
Otherwise, the list contains the standard error (\code{mat_U_error},
\code{mat_F_error}) for both matrices.
}
\description{
Given two matrices of proportions (\code{mat_U} and \code{mat_F}, the growth/survival
matrix and the reproduction matrix respectively) and a sample size, or matrix
of sample sizes, this function calculates the standard error or 95\%
confidence interval for each element of the matrix, depending on the
specified type. These calculations assume that \code{mat_U} is the result of
binomial processes (i.e., the survival (0/1) of individuals), while \code{mat_F}
is the result of Poisson processes (i.e., counts of offspring).
}
\details{
The output is a list containing the error estimates.
}
\examples{

matU <- matrix(c(
  0.1, 0,
  0.2, 0.4
), byrow = TRUE, nrow = 2)
matF <- matrix(c(
  0, 4,
  0., 0.
), byrow = TRUE, nrow = 2)

calculate_errors(mat_U = matU, mat_F = matF, sample_size = 20, type = "CI95")
}
\seealso{
\code{\link[=simulate_mpm]{simulate_mpm()}} which simulates matrices with known values and
sample sizes.
}
\author{
Owen Jones \href{mailto:jones@biology.sdu.dk}{jones@biology.sdu.dk}
}
