% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_models.R
\name{exponential}
\alias{exponential}
\title{Model survival data using the Exponential function}
\usage{
exponential(x, b0)
}
\arguments{
\item{x}{Numeric vector representing age.}

\item{b0}{Numeric value representing the hazard rate parameter.}
}
\value{
A data frame with columns for time (x), hazard (hx), cumulative hazard (Hx), survivorship (Sx) and survival probability within interval (gx).
}
\description{
Model survival data using the Exponential function
}
\examples{
exponential(0:10, 0.01)
}
