% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_models.R
\name{gompertz}
\alias{gompertz}
\title{Model survival data using the Gompertz function}
\usage{
gompertz(x, b0, b1)
}
\arguments{
\item{x}{Numeric vector representing age.}

\item{b0}{Numeric value representing the baseline hazard rate.}

\item{b1}{Numeric value representing the shape parameter.}
}
\value{
A data frame with columns for time (x), hazard (hx), cumulative hazard (Hx), survivorship (Sx) and survival probability within interval (gx).
}
\description{
Model survival data using the Gompertz function
}
\examples{
gompertz(0:10, 0.01, 0.5)
}
