% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_models.R
\name{gompertz_makeham}
\alias{gompertz_makeham}
\title{Model survival data using the Gompertz-Makeham function}
\usage{
gompertz_makeham(x, b0, b1, C)
}
\arguments{
\item{x}{Numeric vector representing age.}

\item{b0}{Numeric value representing the baseline hazard rate.}

\item{b1}{Numeric value representing the Gompertz shape parameter.}

\item{C}{Numeric value representing the Makeham constant.}
}
\value{
A data frame with columns for time (x), hazard (hx), cumulative hazard (Hx), survivorship (Sx) and survival probability within interval (gx).
}
\description{
Model survival data using the Gompertz-Makeham function
}
\examples{
gompertz_makeham(0:10, 0.01, 0.5, 0.1)
}
