% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_leslie_mpm.R
\name{make_leslie_matrix}
\alias{make_leslie_matrix}
\title{Create a Leslie matrix}
\usage{
make_leslie_matrix(survival, fertility, n_stages)
}
\arguments{
\item{survival}{a numeric value representing the survival probability of each stage along the lower off-diagonal of the matrix, with the final value being in the lower-right corner of the matrix. If only one value is provided, this is applied to all survival elements.}

\item{fertility}{a numeric vector of length n_stages representing the fertility rate of each stage. If only one value is provided, this is applied to all fertility elements.}

\item{n_stages}{a numeric value representing the number of stages in the matrix}
}
\value{
A matrix of size n_stages x n_stages representing the Leslie matrix
}
\description{
The function creates a Leslie matrix from inputs of number of stages, fertility (the top row of the matrix), and survival probability (the value in the sub-diagonal).
}
\examples{
make_leslie_matrix(survival = 0.5, fertility = c(0.1, 0.2, 0.3), n_stages = 3)
make_leslie_matrix(survival = c(0.5,0.6,0.7), fertility = c(0.1,0.2,0.3), n_stages = 3)
make_leslie_matrix(survival = seq(0.1,0.7,length.out = 4), fertility = 0.1, n_stages = 4)
make_leslie_matrix(survival = c(0.8,0.3,0.2,0.1,0.05), fertility = 0.2, n_stages = 5)

}
