% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fertility.R
\name{model_fertility}
\alias{model_fertility}
\title{Model fertility with age using set functional forms}
\usage{
model_fertility(age = NULL, params, maturity = 0, model = "logistic")
}
\arguments{
\item{age}{A numeric vector representing age.}

\item{params}{A numeric vector of parameters for the selected model. The
number and meaning of parameters depend on the selected model.}

\item{maturity}{A non-negative numeric value indicating the age at maturity.
Whatever model is used, the fertility is forced to be 0 below the age of
maturity.}

\item{model}{A character string specifying the model to use. Must be one of
"logistic", "step", or "vonbertalanffy".}
}
\value{
A numeric vector representing the computed fertility values.
}
\description{
This function computes fertility based on the logistic, step, von
Bertalanffy, and normal models. The logistic model assumes that fertility
increases sigmoidally with age until a maximum fertility is reached, while
the step model assumes that fertility is zero before the age of maturity and
then remains constant. The von Bertalanffy model assumes that, after
maturity, fertility increases asymptotically with age until a maximum
fertility is reached. For all models, the output ensures that fertility is
zero before the age at maturity.
}
\examples{
# Compute fertility using the logistic model
model_fertility(age = 0:20, params = c(10, 0.5, 8), maturity = 0, model = "logistic")

# Compute fertility using the step model
model_fertility(age = 0:20, params = 10, maturity = 2, model = "step")

# Compute fertility using the von Bertalanffy model
model_fertility(age = 0:20, params = c(10, .3), maturity = 2, model = "vonbertalanffy")

# Compute fertility using the normal model
model_fertility(age = 0:20, params = c(10, 4, 2), maturity = 0, model = "normal")

}
\author{
Owen Jones \href{mailto:jones@biology.sdu.dk}{jones@biology.sdu.dk}
}
