% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmsim-package.R
\docType{package}
\name{mpmsim-package}
\alias{mpmsim}
\alias{mpmsim-package}
\title{mpmsim: Simulation of Matrix Population Models with Defined Life History Characteristics}
\description{
Allows users to simulate matrix population models with particular characteristics based on aspects of life history such as mortality trajectories and fertility trajectories. Also allows the exploration of sampling error due to small sample size.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonesor/mpmsim}
  \item Report bugs at \url{https://github.com/jonesor/mpmsim/issues}
}

}
\author{
\strong{Maintainer}: Owen Jones \email{jones@biology.sdu.dk} (\href{https://orcid.org/0000-0001-5720-4686}{ORCID})

}
\keyword{internal}
