% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpmsim-package.R
\docType{package}
\name{mpmsim}
\alias{mpmsim}
\title{mpmsim}
\description{
Simulates Matrix Population Models with Particular Life History
Characteristics
}
\author{
Owen Jones \href{mailto:jones@biology.sdu.dk}{jones@biology.sdu.dk}
}
