% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_leslie_set.R
\name{rand_leslie_set}
\alias{rand_leslie_set}
\title{Generate a set of random Leslie Matrix Population Models}
\usage{
rand_leslie_set(
  n_models = 5,
  mortality_model = "gompertz",
  fertility_model = "step",
  mortality_params,
  fertility_params,
  fertility_maturity_params,
  dist_type = "uniform",
  output = "mpm",
  split = TRUE,
  by_type = TRUE,
  as_compadre = TRUE
)
}
\arguments{
\item{n_models}{An integer indicating the number of MPMs to generate.}

\item{mortality_model}{A character string specifying the name of the
mortality model to be used. Options are \code{Gompertz}, \code{GompertzMakeham},
\code{Exponential}, \code{Siler}, \code{Weibull}, and \code{WeibullMakeham}. See \code{model_mortality}.}

\item{fertility_model}{A character string specifying the name of the
fertility model to be used. Options are \code{logistic}, \code{step},
\code{vonBertalanffy}, \code{normal} and \code{hadwiger.} See \code{?model_fertility}.}

\item{mortality_params}{A two-column dataframe with a number of rows equal to
the number of parameters in the mortality model. The required order of the
parameters depends on the selected \code{mortality_model} (see
\code{?model_mortality}):
\itemize{
\item For \code{Gompertz} and \code{Weibull}: \code{b_0}, \code{b_1}
\item For \code{GompertzMakeham} and \code{WeibullMakeham}: \code{b_0}, \code{b_1}, \code{C}
\item For \code{Exponential}: \code{C}
\item For \code{Siler}: \code{a_0}, \code{a_1}, \code{C}, \code{b_0}, \code{b_1}
If \code{dist_type} is \code{uniform} these rows represent the lower and upper limits
of the random uniform distribution from which the parameters are drawn. If
\code{dist_type} is \code{normal}, the columns represent the mean and standard
deviation of a random normal distribution from which the parameter values
are drawn.
}}

\item{fertility_params}{A two-column dataframe with a number of rows equal to
the number of parameters in the fertility model. The required order of the
parameters depends on the selected \code{fertility_model} (see
\code{?model_mortality}):
\itemize{
\item For \code{logistic}: \code{A}, \code{k}, \code{x_m}
\item For \code{step}: \code{A}
\item For \code{vonBertalanffy}: \code{A}, \code{k}
\item For \code{normal}: \code{A}, \code{mu}, \code{sd}
\item For \code{hadwiger}: \code{a}, \code{b}, \code{C}
If \code{dist_type} is \code{uniform} these rows represent the lower and upper limits
of the random uniform distribution from which the parameters are drawn. If
\code{dist_type} is \code{normal}, the columns represent the mean and standard
deviation of a random normal distribution from which the parameter values
are drawn.
}}

\item{fertility_maturity_params}{A vector with two elements defining the
distribution from which age at maturity is drawn for the models. The models
will coerce fertility to be zero before this point. If \code{dist_type} is
\code{uniform} these values represent the lower and upper limits of the random
uniform distribution from which the parameters are drawn. If \code{dist_type} is
\code{normal}, the values represent the mean and standard deviation of a random
normal distribution from which the parameter values are drawn.}

\item{dist_type}{A character string specifying the type of distribution to
draw parameters from. Default is \code{uniform}. Supported types are \code{uniform}
and \code{normal}.}

\item{output}{Either \code{mpm} or \code{lifetable}, to output either as matrix
population models or as life tables respectively. Default is \code{mpm}.}

\item{split}{A logical indicating whether to split MPMs into submatrices.
Default is TRUE.Ignored if output is \code{lifetable}.}

\item{by_type}{A logical value indicating how the matrices should be
returned. If \code{TRUE}, the function returns a list containing three separate
lists for each matrix type (A, U, and F). If \code{FALSE}, the function returns
a list where each element contains three matrices representing a single
model grouped together (A, U, and F, where A = U + F). When \code{split} is
\code{FALSE}, \code{by_type} is automatically set to \code{FALSE}. Defaults to \code{TRUE}.
Ignored if output is \code{lifetable}.}

\item{as_compadre}{A logical indicating whether the matrices should be
returned as a \code{CompadreDB} object. Default is \code{TRUE}. If \code{FALSE}, the
function returns a list. Ignored if output is \code{lifetable}.}
}
\value{
If output is \code{mpm}, this function returns a \code{compadreDB} object or
list containing MPMs generated using the specified model with parameters
drawn from random uniform or normal distributions. The format of the output
MPMs depends on the arguments \code{by_type}, \code{split} and \code{as_compadre}. If
output is \code{lifetable}, the function returns a list of life tables.
}
\description{
Generates a set of Leslie matrix population models (MPMs) based on defined
mortality and fertility models, and using model parameters randomly drawn
from specified distributions.
}
\examples{

mortParams <- data.frame(minVal = c(0, 0.01,0.1), maxVal = c(0.14, 0.15, 0.1))
fertParams <- data.frame(minVal = c(10, 0.5, 8), maxVal = c(11, 0.9, 10))
maturityParam <- c(0, 0)

rand_leslie_set(n_models = 5,
                mortality_model = "GompertzMakeham",
                fertility_model = "normal",
                mortality_params = mortParams,
                fertility_params = fertParams,
                fertility_maturity_params = maturityParam,
                dist_type="uniform",
                output = "mpm")

}
\seealso{
Other Leslie matrices: 
\code{\link{make_leslie_mpm}()},
\code{\link{reorganise_matrices}()}
}
\author{
Owen Jones \href{mailto:jones@biology.sdu.dk}{jones@biology.sdu.dk}
}
\concept{Leslie matrices}
