% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_models.R
\name{siler}
\alias{siler}
\title{Compute Siler Model}
\usage{
siler(x, a0, a1, C, b0, b1)
}
\arguments{
\item{x}{a numeric vector representing time}

\item{a0}{a numeric value representing the first parameter of the hazard function}

\item{a1}{a numeric value representing the second parameter of the first exponential term of the hazard function}

\item{C}{a numeric value representing the constant term of the hazard function}

\item{b0}{a numeric value representing the third parameter of the hazard function}

\item{b1}{a numeric value representing the fourth parameter of the second exponential term of the hazard function}
}
\value{
A data frame with columns for time (x), hazard (hx), cumulative hazard (Hx), survivorship (Sx) and survival probability within interval (gx).
}
\description{
Compute Siler Model
}
\examples{
siler(0:10, a0 = 0.5, a1 = 0.2, C = 0.1, b0 = 0.1, b1 = 0.2)

}
