% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mpm_set.R
\name{generate_mpm_set}
\alias{generate_mpm_set}
\title{Generate lists of Lefkovitch matrix population models (MPMs) based on life
history archetypes}
\usage{
generate_mpm_set(
  n = 10,
  lower_lambda = 0.9,
  upper_lambda = 1.1,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = TRUE,
  by_type = TRUE,
  max_surv = 0.99,
  constraint = NULL,
  attempts = 1000
)
}
\arguments{
\item{n}{The number of MPMs to generate. Default is 10.}

\item{lower_lambda}{The lower bound for acceptable lambda values. Default is
0.9.}

\item{upper_lambda}{The upper bound for acceptable lambda values. Default is
1.1.}

\item{n_stages}{The number of stages for the MPMs. Default is 3.}

\item{archetype}{The archetype of the MPMs. Default is 1.}

\item{fecundity}{A vector of fecundities for the MPMs. Default is 1.5.}

\item{split}{A logical indicating whether to split into submatrices. Default is
TRUE.}

\item{by_type}{A logical indicating whether the matrices should be returned
in a list by type (A, U, F, C). If split is \code{FALSE}, then \code{by_type} must
also be \code{FALSE}.}

\item{max_surv}{The maximum acceptable survival value. Defaults to 0.99. This
is only used if \code{split = TRUE}.}

\item{constraint}{An optional list specifying (1) a function that outputs a metric
derived from an A matrix and (2) acceptable bounds that are used to
constrain matrix outputs. E.g. \code{list("generation.time", c(5,10))} to
specify that \code{generate_mpm_set} should only output MPMs with a generation
time between 5 and 10. Default is \code{NULL}.}

\item{attempts}{an integer indicating the number of attempts To be made when
simulating matrix model. The default is 1000. If it takes more than 1000
attempts to make a matrix that satisfies the conditions set by the other
arguments, then a warning is produced.}
}
\value{
A list of MPMs that meet the specified criteria.
}
\description{
This function generates a list of \code{n} MPMs according to the specified
criteria. Criteria include the \code{archetype}, and the acceptable lambda values.
The function attempts to find matrices that fulfill the lambda criteria,
discarding unacceptable matrices. If it takes more than 1000 attempts to find
a suitable matrix model, then an error is produced.
}
\examples{
generate_mpm_set(
  n = 10, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE
)

library(popbio)
generate_mpm_set(
  n = 1, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE,
  constraint = list("generation.time", c(40, 50))
)

}
\seealso{
\code{\link[=random_mpm]{random_mpm()}} which this function is essentially a wrapper for.

Other simulation: 
\code{\link{random_mpm}()}
}
\author{
Owen Jones \href{mailto:jones@biology.sdu.dk}{jones@biology.sdu.dk}
}
\concept{simulation}
