% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mpm_set.R
\name{generate_mpm_set}
\alias{generate_mpm_set}
\title{Generate sets of MPMs}
\usage{
generate_mpm_set(
  n = 10,
  lower_lambda = 0.9,
  upper_lambda = 1.1,
  nStage = 3,
  archetype = 1,
  Fec = 1.5,
  split = FALSE
)
}
\arguments{
\item{n}{The number of MPMs to generate. Default is 10.}

\item{lower_lambda}{The lower bound for acceptable lambda values. Default is 0.9.}

\item{upper_lambda}{The upper bound for acceptable lambda values. Default is 1.1.}

\item{nStage}{The number of stages for the MPMs. Default is 3.}

\item{archetype}{The archetype of the MPMs. Default is 1.}

\item{Fec}{A vector of fecundities for the MPMs. Default is 1.5.}

\item{split}{A logical indicating whether to split matrices. Default is FALSE.}
}
\value{
A list of MPMs that meet the specified criteria.
}
\description{
This function generates a set of n MPMs according to the specified criteria.
Criteria include the archetype, and the acceptable lambda values.
The function attempts to find matrices that fulfil the lambda criteria, discarding unacceptable matrices. If it takes more than 1000 attempts to find a suitable matrix model, then an error is produced.
}
\examples{
generate_mpm_set(
  n = 10, lower_lambda = 0.9, upper_lambda = 1.1,
  nStage = 5, Fec = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE
)

}
\author{
Owen Jones \href{mailto:jones@biology.sdu.dk}{jones@biology.sdu.dk}
}
