% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_mpm_set.R
\name{generate_mpm_set}
\alias{generate_mpm_set}
\title{Generate lists of matrix population models (MPMs) based on life history
archetypes}
\usage{
generate_mpm_set(
  n = 10,
  lower_lambda = 0.9,
  upper_lambda = 1.1,
  n_stages = 3,
  archetype = 1,
  fecundity = 1.5,
  split = FALSE
)
}
\arguments{
\item{n}{The number of MPMs to generate. Default is 10.}

\item{lower_lambda}{The lower bound for acceptable lambda values. Default is
0.9.}

\item{upper_lambda}{The upper bound for acceptable lambda values. Default is
1.1.}

\item{n_stages}{The number of stages for the MPMs. Default is 3.}

\item{archetype}{The archetype of the MPMs. Default is 1.}

\item{fecundity}{A vector of fecundities for the MPMs. Default is 1.5.}

\item{split}{A logical indicating whether to split matrices. Default is
FALSE.}
}
\value{
A list of MPMs that meet the specified criteria.
}
\description{
This function generates a list of \code{n} MPMs according to the specified
criteria. Criteria include the \code{archetype}, and the acceptable lambda values.
The function attempts to find matrices that fulfill the lambda criteria,
discarding unacceptable matrices. If it takes more than 1000 attempts to find
a suitable matrix model, then an error is produced.
}
\examples{
generate_mpm_set(
  n = 10, lower_lambda = 0.9, upper_lambda = 1.1,
  n_stages = 5, fecundity = c(0, 0, 4, 8, 10), archetype = 4, split = TRUE
)

}
\seealso{
\code{\link[=random_mpm]{random_mpm()}} which this function is essentially a wrapper for.
}
\author{
Owen Jones \href{mailto:jones@biology.sdu.dk}{jones@biology.sdu.dk}
}
