% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mortality_models.R
\name{model_survival}
\alias{model_survival}
\title{Model survival data using a mortality model}
\usage{
model_survival(x, params, model)
}
\arguments{
\item{x}{Numeric vector representing age.}

\item{params}{Numeric vector representing the parameters of the mortality
model.}

\item{model}{Mortality model: \code{Gompertz}, \code{GompertzMakeham}, \code{Exponential},
\code{Siler}.}
}
\value{
A data frame with columns for time (\code{x}), hazard (\code{hx}), cumulative
hazard (\code{Hx}), survivorship (\code{Sx}) and survival probability within interval
(\code{gx}).
}
\description{
Model survival data using a mortality model
}
\details{
The required parameters varies depending on the mortality model. The
parameters are provided as a vector. For Gompertz, the parameters are b0,
b1. For Gompertz-Makeham the parameters are b0, b1 and C. For Exponential,
the parameter is C. For Siler, the parameters are a0,a1, C, b0 and b1.
}
\examples{
model_survival(0:10, c(0.1, 0.2), "Gompertz")
model_survival(0:10, c(0.1, 0.2, 0.1), "GompertzMakeham")
model_survival(0:10, 0.2, "Exponential")
model_survival(0:10, c(0.1, 0.2, 0.1, 0.1, 0.2), "Siler")
}
