% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_mpm.R
\name{simulate_mpm}
\alias{simulate_mpm}
\title{Simulate matrix population model}
\usage{
simulate_mpm(matU, matF, sample_size, split = TRUE)
}
\arguments{
\item{matU}{matrix of survival probabilities}

\item{matF}{matrix of mean fecundity values}

\item{sample_size}{sample size}

\item{split}{logical, whether to split the output into survival and fecundity matrices or not}
}
\value{
list of matrices of survival and fecundity if split = TRUE, otherwise a single matrix of the sum of survival and fecundity
}
\description{
Simulate matrix population model
}
\examples{
mats <- make_leslie_matrix(
  survival = c(0.1, 0.2, 0.5),
  fertility = c(0, 1.2, 2.4),
  n_stages = 3, split = TRUE
)
ssMat <- matrix(10, nrow = 3, ncol = 3)

simulate_mpm(
  matU = mats$matU, matF = mats$matF,
  sample_size = ssMat, split = TRUE
)
}
